Migrate Why to styled-component (#340)
